Minor utils package refactoring.
