+++ src/main/java/com/github/evillootlye/caves/caverns/CavesAging.java
@@ -3,12 +3,12 @@ package com.github.evillootlye.caves.caverns;
 import com.github.evillootlye.caves.configuration.Configurable;
 import com.github.evillootlye.caves.ticks.TickLevel;
 import com.github.evillootlye.caves.ticks.Tickable;
-import com.github.evillootlye.caves.utils.Materials;
-import com.github.evillootlye.caves.utils.Utils;
-import com.github.evillootlye.caves.utils.bounds.Bound;
-import com.github.evillootlye.caves.utils.bounds.DualBound;
-import com.github.evillootlye.caves.utils.bounds.SingularBound;
-import com.github.evillootlye.caves.utils.random.Rnd;
+import com.github.evillootlye.caves.util.Materials;
+import com.github.evillootlye.caves.util.Utils;
+import com.github.evillootlye.caves.util.bounds.Bound;
+import com.github.evillootlye.caves.util.bounds.DualBound;
+import com.github.evillootlye.caves.util.bounds.SingularBound;
+import com.github.evillootlye.caves.util.random.Rnd;
 import org.bukkit.Bukkit;
 import org.bukkit.Chunk;
 import org.bukkit.Material;
