+++ src/main/java/com/github/evillootlye/caves/mobs/MobsManager.java
@@ -5,10 +5,10 @@ import com.github.evillootlye.caves.DangerousCaves;
 import com.github.evillootlye.caves.configuration.Configurable;
 import com.github.evillootlye.caves.ticks.TickLevel;
 import com.github.evillootlye.caves.ticks.Tickable;
-import com.github.evillootlye.caves.utils.Locations;
-import com.github.evillootlye.caves.utils.Utils;
-import com.github.evillootlye.caves.utils.random.AliasMethod;
-import com.github.evillootlye.caves.utils.random.Rnd;
+import com.github.evillootlye.caves.util.Locations;
+import com.github.evillootlye.caves.util.Utils;
+import com.github.evillootlye.caves.util.random.AliasMethod;
+import com.github.evillootlye.caves.util.random.Rnd;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.World;
