+++ src/main/java/com/github/evillootlye/caves/util/Locations.java
@@ -1,4 +1,4 @@
-package com.github.evillootlye.caves.utils;
+package com.github.evillootlye.caves.util;
 
 import org.bukkit.Location;
 import org.bukkit.World;
