+++ src/main/java/com/github/evillootlye/caves/util/bounds/Bound.java
@@ -1,4 +1,4 @@
-package com.github.evillootlye.caves.utils.bounds;
+package com.github.evillootlye.caves.util.bounds;
 
 public interface Bound {
     boolean isInside(int x, int z);
